requirements.txt
django-model-utils>=2.0
# Additional requirements go here
requests
lxml
python-dateutil
pendulum
anyjson
simplejson
swapper
